feat: Add user registration functionality
Add user registration functionality to the auth module. This allows users to create an account by providing their name, email, and password. The registration process includes hashing the password and creating a new user record in the database. If the registration is successful, the user is assigned to a default team and becomes the owner of that team. If the registration fails, an appropriate error message is returned. This feature enhances the application by enabling users to sign up and start using the platform.